core-customize/hybris/bin/modules/data-hub-adapter/datahubbackoffice/backoffice/src/de/hybris/platform/datahubbackoffice/service/datahub/impl/UserContextImpl.java
/*
 * Copyright (c) 2022 SAP SE or an SAP affiliate company. All rights reserved.
 */
package de.hybris.platform.datahubbackoffice.service.datahub.impl;

import de.hybris.platform.core.model.user.UserGroupModel;
import de.hybris.platform.core.model.user.UserModel;
import de.hybris.platform.datahubbackoffice.ApplicationBeans;
import de.hybris.platform.datahubbackoffice.service.datahub.UserContext;
import de.hybris.platform.servicelayer.user.UserService;

import java.io.Serial;
import java.io.Serializable;

/**
 * Stateful service for retrieving information about the user in the context of a web request.
 */
public class UserContextImpl implements UserContext, Serializable
{
	@Serial
	private static final long serialVersionUID = 1734800045352907935L;

	private transient UserService userService;
	private UserModel currentUser;
	private UserGroupModel datahubAdminGroup;

	@Override
	public UserModel getCurrentUser()
	{
		if (currentUser == null)
		{
			currentUser = getUserService().getCurrentUser();
		}
		return currentUser;
	}

	@Override
	public boolean isUserDataHubAdmin()
	{
		return isMemberOf(getDataHubAdminGroup());
	}

	@Override
	public boolean isMemberOf(final String userGroup)
	{
		return isMemberOf(getUserGroup(userGroup));
	}

	private boolean isMemberOf(final UserGroupModel group)
	{
		return userService.isMemberOfGroup(getCurrentUser(), group);
	}

	private UserGroupModel getDataHubAdminGroup()
	{
		if (datahubAdminGroup == null)
		{
			datahubAdminGroup = getUserGroup("datahubadmingroup");
		}
		return datahubAdminGroup;
	}

	private UserGroupModel getUserGroup(final String groupId)
	{
		return userService.getUserGroupForUID(groupId);
	}

	public void setUserService(final UserService userService)
	{
		this.userService = userService;
	}

	private UserService getUserService()
	{
		if (userService == null)
		{
			userService = ApplicationBeans.getBean("userService", UserService.class);
		}
		return userService;
	}
}
